pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "three-d-genome-hub"
version = "0.1.0"
description = "AI-ready knowledge base for 3D genome education and research"
authors = [
  { name = "3DGenomeHub Maintainers", email = "maintainers@example.com" }
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "typer>=0.12",
  "pydantic>=1.10",
  "PyYAML>=6.0",
  "scikit-learn>=1.4",
  "pandas>=2.2",
  "SQLAlchemy>=2.0",
  "python-dotenv>=1.0",
  "joblib>=1.3"
]

[project.scripts]
three-d-genome-hub = "three_d_genome_hub.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
